+++ R/99_imports.R
@@ -1,8 +1,10 @@
-#' @importFrom dplyr %>% 
 #' @importFrom rlang !! := .data
 #' @importFrom  grid grid.raster grid.polygon grid.lines grid.points grid.text gpar unit
 #'
 #'
 NULL
 
+#' @importFrom magrittr %>%
+#' @export
+magrittr::`%>%`