+++ R/40_kaarten.R
@@ -3,6 +3,7 @@
 #' Deze functie genereert een basiskaart in leaflet. Aan deze kaart kunnen vervolgens vrij elementen worden toegevoegd
 #'
 #' @param data Een optioneel dataframe. Zie ook het data-argument in de functie [leaflet::leaflet()]
+#' @param type Het type kaart. Opties zijn `"osm"` (default) en `"cartolight"`
 #'
 #' @return Een leaflet kaart
 #' @export
@@ -11,10 +12,13 @@
 #' 
 #' basiskaart()
 #' 
-basiskaart <- function(data = NULL) {
+basiskaart <- function(data = NULL, type = c("osm", "cartolight")) {
+  type <- rlang::arg_match(type, c("osm", "cartolight"))
+  
   leaflet::leaflet(data) %>% 
-    leaflet::addProviderTiles(leaflet::providers$OpenStreetMap, group = "Kaart") %>% 
-    leaflet::addProviderTiles(leaflet::providers$Esri.WorldImagery, group = "Luchtfoto") %>% 
+    {if (type == "osm") {leaflet::addProviderTiles(. ,"OpenStreetMap", group = "Kaart") } else {.}} %>%
+    {if (type == "cartolight") {leaflet::addProviderTiles(., "CartoDB.Positron", group = "Kaart") } else {.}} %>%
+    leaflet::addProviderTiles("Esri.WorldImagery", group = "Luchtfoto") %>% 
     leaflet::addLayersControl( baseGroups = c("Kaart", "Luchtfoto"), 
                       options = leaflet::layersControlOptions(collapsed = FALSE),
                       position = "topleft")