+++ NEWS.md
@@ -1,6 +1,8 @@
 # HHSKwkl 0.2.1
 
 - Kleuren van hhskthema iets verbeterd en grafiek_basis iets gewijzigd.
+- Interne functies voor testdata toegevoegd.
+- Import_fys_chem sorteert op mp, parnr, datum en waarde
 
 # HHSKwkl 0.2.0