+++ NEWS.md
@@ -3,7 +3,8 @@
 - Kleuren van hhskthema iets verbeterd en grafiek_basis iets gewijzigd.
 - Interne functies voor testdata toegevoegd.
 - Import_fys_chem sorteert op mp, parnr, datum en waarde
-- deprecated the my_c function
+- deprecated the `my_c` function
+- `frac()` en `perc()` toegevoegd om snel fracties in summarise te kunnen maken.
 
 # HHSKwkl 0.2.0