+++ NEWS.md
@@ -1,3 +1,5 @@
+# HHSKwkl (development version)
+
 # HHSKwkl 0.2.1
 
 - Kleuren van hhskthema iets verbeterd en grafiek_basis iets gewijzigd.