+++ NEWS.md
@@ -1,4 +1,4 @@
-# HHSKwkl (development version)
+# HHSKwkl 0.3.0
 
 - `import_normen_rivm()` bijgewerkt met meer info
 - `hhskthema_kaart()` toegevoegd